Be/dev

See merge request s12-blockchain-nft-sub1/S12P21D210!114

+++ Backend/src/main/java/ssafy/d210/backend/entity/UserLecture.java
@@ -33,7 +33,7 @@ public class UserLecture {
     private LocalDate certificateDate;
 
     @ColumnDefault("0")
-    private int recentLectureId;
+    private Long recentLectureId;
 
     private String qrCode;
